Entregar/requirements.txt
azure-functions
pyodbc
pandas
sklearn
joblib